Potential COVID19 cases reported through NHS pathways

This dataset contains daily numbers of reports on potential COVID-19 cases reported in England through the NHS 111 calls, 999 calls, and 111-online systems. The present dataset was last updated on 21 September 2020. See example for a command line allowing to download the latest version.

covid19_england_nhscalls_2020

Format

A `data.frame` containing:

site_type

the system through which data were reported: 111/999 calls, or 111-online

date

the data of reporting

sex

the gender of the patient

age

the age of the patient, in years

ccg_code

NHS code for the Clinical Commissioning Groups (CCGs) (finer geographic unit)

ccg_name

name of the Clinical Commissioning Groups (CCGs) (smaller spatial unit)

count

number of potential COVID-19 cases reported

postcode

the postcode of the CCG

nhs_region

the NHS region (larger geographic unit)

day

the date as the number of days since the first reporting day

weekday

the day of the week, broken down into: weekend, Monday, and the rest of the week; this is used for modelling reporting effects

Source

Data is available at https://digital.nhs.uk/dashboards/nhs-pathways; this precise dataset adds some cleaning and additional informaion (on NHS regions) and is taken from Quentin Leclerc's github repository: https://github.com/qleclerc/nhs_pathways_report

Author

National Health Services (NHS) for England. Additional data and cleaning by Quentin Leclerc.

Examples

if (FALSE) {
# These commands will download the latest version of the data:
library(tidyverse)

# download data
pathways <- tempfile()
url <- paste0("https://github.com/qleclerc/nhs_pathways_report/",
              "raw/master/data/rds/pathways_latest.rds")
download.file(url, pathways)
pathways <- readRDS(pathways)
head(pathways)

}