Influenza A H7N9 outbreak in China

These data comprise of 136 cases of influenza A H7N9 in China, analysed by Kucharski et al. (2014).

fluH7N9.china.2013

Format

A data frame with 136 rows and 8 columns

Source

http://datadryad.org/resource/doi:10.5061/dryad.2g43n/1

References

A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014. Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data. PLOS Currents Outbreaks. Mar 7, edition 1. doi: 10.1371/currents.outbreaks.e1473d9bfc99d080ca242139a06c455f.

A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014. Data from: Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data. Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.2g43n.

Examples

## show first few cases
head(fluH7N9.china.2013)

#>   case.ID date.of.onset date.of.hospitalisation date.of.outcome outcome gender
#> 1       1    2013-02-19                    <NA>      2013-03-04   Death      m
#> 2       2    2013-02-27              2013-03-03      2013-03-10   Death      m
#> 3       3    2013-03-09              2013-03-19      2013-04-09   Death      f
#> 4       4    2013-03-19              2013-03-27            <NA>    <NA>      f
#> 5       5    2013-03-19              2013-03-30      2013-05-15 Recover      f
#> 6       6    2013-03-21              2013-03-28      2013-04-26   Death      f
#>   age province
#> 1  87 Shanghai
#> 2  27 Shanghai
#> 3  35    Anhui
#> 4  45  Jiangsu
#> 5  48  Jiangsu
#> 6  32  Jiangsu