fluH7N9_china_2013 {outbreaks}	R Documentation

Influenza A H7N9 in China, 2013

Description

These data comprise of 136 cases of influenza A H7N9 in China, analysed by Kucharski et al. (2014).

Usage

fluH7N9_china_2013

Format

A data frame with 136 rows and 8 columns

Author(s)

Data collated by Adam Kucharski et al. from ProMed, WHO, FluTrackers, news reports and research articles. Transfer to R and documentation by Simon Frost (sdwfrost@gmail.com).

Source

http://datadryad.org/resource/doi:10.5061/dryad.2g43n/1

References

A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014. Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data. PLOS Currents Outbreaks. Mar 7, edition 1. doi: 10.1371/currents.outbreaks.e1473d9bfc99d080ca242139a06c455f.

A. Kucharski, H. Mills, A. Pinsent, C. Fraser, M. Van Kerkhove, C. A. Donnelly, and S. Riley. 2014. Data from: Distinguishing between reservoir exposure and human-to-human transmission for emerging pathogens using case onset data. Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.2g43n.

Examples

## show first few cases
head(fluH7N9_china_2013)

[Package outbreaks version 1.5.0 Index]