Influenza in a boarding school in England, 1978

These data comprise of a time series of influenza cases in a boarding school in England. The original data were available only in a figure with some additional data in the main text; hence, the exact numbers vary depending on the source. These data are from Chapter 9 of De Vries et al. (1996).

influenza_england_1978_school

Format

A data frame with 14 rows and 3 columns

date

Date

in_bed

Number in bed

convalescent

Number convalescing

Source

http://www.ncbi.nlm.nih.gov/pmc/articles/PMC1603269/pdf/brmedj00115-0064.pdf

Details

The index case was infected by 1978-01-10, and had febrile illness from 1978-01-15 to 1978-01-18. 512 boys out of 763 became ill.

References

Anonymous. 1978. Influenza in a boarding school. British Medical Journal 1:578.

G. De Vries, T. Hillen, M. Lewis, J. Mueller, and B. Schoenfisch. 2006. A Course in Mathematical Biology: Quantitative Modeling with Mathematical and Computational Methods. Society for Industrial and Applied Mathematics.

Examples

## show first few cases
head(influenza_england_1978_school)

#>         date in_bed convalescent
#> 1 1978-01-22      3            0
#> 2 1978-01-23      8            0
#> 3 1978-01-24     26            0
#> 4 1978-01-25     76            0
#> 5 1978-01-26    225            9
#> 6 1978-01-27    298           17