Middle East respiratory syndrome in South Korea, 2015

These datasets correspond to the initial information collected by the Epidemic Intelligence group at European Centre for Disease Prevention and Control (ECDC) during the first weeks of the outbreak of Middle East respiratory syndrome (MERS-CoV) outbreak (South Korea) in 2015. The data were used to follow the daily evolution of this outbreak using public information available.

mers_korea_2015

Format

A list of two dataframes:

$linelist: A dataframe of MERS-CoV cases and their attributes

id: Unique identifier

age: Age

age_class: Age using 10-year groups

sex: Sex

place_infect: Probable region of infection

reporting_ctry: Country reporting the case

loc_hosp: Local hospital name where the case was hospitalized

dt_onset: Date of onset of symptoms

dt_report: Date of reporting

week_report: Week number of date of reporting

dt_start_exp: Date of first probable exposure to another MERS Co-V case

dt_end_exp: Date of last probable exposure to another MERS Co-V case

dt_diag: Date of MERS Co-V diagnosis

outcome: Outcome (alive or dead)

dt_death: Date of death

$contacts: A dataframe describing the relationship between MERS Co-V cases

from: Unique identifier of the probably source patient

to: Unique identifier of the secondary case

exposure: Probable place of exposure

diff_dt_onset: Time in days between two successive cases

Details

This dataset is meant for teaching purposes; it represents neither the final outbreak investigation results nor a consolidated and complete description of the transmission chain.

References

More information on the intial stage of the outbreak in the following reference: Penttinen PM, Kaasik-Aaslav K, Friaux A, Donachie A, Sudre B, Amato-Gauci AJ, Memish ZA, Coulombier D. Taking stock of the first 133 MERS coronavirus cases globally--Is the epidemic changing? Euro Surveill. 2013 Sep 26;18(39). pii: 20596. PubMed PMID: 24094061.

Examples

## show the line list describing MERS Co-V cases and their attributes
head(mers_korea_2015$linelist)

#>     id age age_class sex        place_infect reporting_ctry
#> 1 SK_1  68     60-69   M         Middle East    South Korea
#> 2 SK_2  63     60-69   F Outside Middle East    South Korea
#> 3 SK_3  76     70-79   M Outside Middle East    South Korea
#> 4 SK_4  46     40-49   F Outside Middle East    South Korea
#> 5 SK_5  50     50-59   M Outside Middle East    South Korea
#> 6 SK_6  71     70-79   M Outside Middle East    South Korea
#>                                              loc_hosp   dt_onset  dt_report
#> 1 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-11 2015-05-19
#> 2 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-18 2015-05-20
#> 3 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-20 2015-05-20
#> 4 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-25 2015-05-26
#> 5                           365 Yeollin Clinic, Seoul 2015-05-25 2015-05-27
#> 6 Pyeongtaek St. Mary, Hospital, Pyeongtaek, Gyeonggi 2015-05-24 2015-05-28
#>   week_report dt_start_exp dt_end_exp    dt_diag outcome   dt_death
#> 1     2015_21   2015-04-18 2015-05-04 2015-05-20   Alive       <NA>
#> 2     2015_21   2015-05-15 2015-05-20 2015-05-20   Alive       <NA>
#> 3     2015_21   2015-05-16 2015-05-16 2015-05-21    Dead 2015-06-04
#> 4     2015_22   2015-05-16 2015-05-20 2015-05-26   Alive       <NA>
#> 5     2015_22   2015-05-17 2015-05-17 2015-05-26   Alive       <NA>
#> 6     2015_22   2015-05-15 2015-05-17 2015-05-28    Dead 2015-06-01

## show the relationships between MERS Co-V cases
head(mers_korea_2015$contacts)

#>    from     to       exposure diff_dt_onset
#> 1 SK_14 SK_113 Emergency room            10
#> 2 SK_14 SK_116 Emergency room            13
#> 3 SK_14  SK_41 Emergency room            14
#> 4 SK_14 SK_112 Emergency room            14
#> 5 SK_14 SK_100 Emergency room            15
#> 6 SK_14 SK_114 Emergency room            15