rabies_car_2003 {outbreaks}	R Documentation

Dog Rabies in Central African Republic, 2003-2012

Description

These data document a dog rabies epidemic from 2003 to 2012 in Bangui, Central African Republic, and its surroundings. Data comprise dates and locations of the cases, as well as viral sequences of the pathogen for most cases.

Usage

rabies_car_2003

Format

A list comprising a data.frame ($linelist) and a DNAbin matrix ($dna). $linelist contains the following variables:

$index: numeric identifier of the case

$date: date of case reporting

$latitude: the latitude of the collection point

$longitude: the longitude of the collection point

$has_dna: a logical indicating of the case has a matching pathogen sequence in $dna

$dna is a DNAbin matrix whose labels are to be matched against $linelist$index.

Author(s)

Data from Transfer to R and documentation by Thibaut Jombart.

Source

The data were provided by the Institut Pasteur de Bangui, Bangui, République Centrafricaine, and the Institut Pasteur, Unit Lyssavirus Dynamics and Host Adaptation, WHO Collaborating Centre for Reference and Research on Rabies, Paris, France.

References

Cori et al. (submitted) A graph-based evidence synthesis approach to detecting outbreak clusters: an application to dog rabies.

Examples

if (require(incidence) && require(ape)) {
  i <- incidence(rabies_car_2003$linelist$date, 28L)
  plot(i)

  tre <- nj(dist.dna(rabies_car_2003$dna))
  plot(tre, main = "Neighbour-Joining tree")

}

[Package outbreaks version 1.5.0 Index]