s_enteritidis_pt59 {outbreaks}	R Documentation

Salmonella Enteritidis PT59 outbreak

Description

This dataset covers a food-borne outbreak of Salmonella Enteritidis PT59 investigated by Public Health England. The data includes a distribution network, and genetic clusters of bacteria isolated in contaminated patients. The data is anonymised: identifiers of the food distribution network have been hashed. The object s_enteritidis_pt59 is a list containing:

Usage

s_enteritidis_pt59

Format

An object of class list of length 2.

Details

$graph: a data.frame containing with two columns from and to specifying (directed) edges of the food distribution network, alongside reported cases (terminal branches, or 'tips').

$cluster: a factor indicating the genetic cluster of named tips.

Author(s)

Thomas Inns Thomas.Inns@phe.gov.uk, Philip Ashton, Tim Dallman, Roberto Vivancos. Hashing and port to R by Thibaut Jombart.

Source

Multi-agency Salmonella Enteritidis PT59 Outbreak Control Team, chaired by Public Health England

[Package outbreaks version 1.5.0 Index]