SARS-CoV-2 World Health Organization Situation Reports 2019 Outbreak (COVID-19)

These data are transcribed from the WHO Situation Reports on the COVID-19 outbreak (SARS-CoV-2). Data was not available for all variables in all reports. For full details, see the original Situation Reports as published by WHO. Data were manually transcribed and errors are possible.

sarscov2_who_2019

Format

A data frame where each row represents a new Situation Report

Source

World Health Organization (2020)

References

World Health Organization. 2020. <https://www.who.int/emergencies/diseases/novel-coronavirus-2019/situation-reports>

Author

Data from World Health Organization (WHO), published as Sitation Reports. Transfer to R and documentation by Eric Brown.

Examples

## show global cases
sarscov2_who_2019$cases_global

#>  [1]     282     314     581     846    1320    2014    2798    4593    6065
#> [10]    7818    9826   11953   14557   17391   20630   24554   28276   31481
#> [19]   34886   37558   40554   43103   45171   46997   49053   50580   51857
#> [28]   71429   73332   75204   75748   76769   77794   78811   79331   80239
#> [37]   81109   82294   83652   85403   87137   88948   90870   93090   95333
#> [46]   98192  101927  105586  109578  113702  118326  125048  132758  142539
#> [55]  153517  167511  179112  191127  209839  234073  266073  292142  332930
#> [64]  372757  414179  462684  509164  571678  634835  693224  750890  823626
#> [73]  896450  972303 1051635 1133758 1210956 1279722 1353361 1436198 1521252
#> [82] 1610909 1696588 1773084 1844863 1914916 1991562 2074529 2160207 2241778
#> [91] 2314621 2397216