Zika in San Andres, Colombia, 2015

These data describe the daily incidence of Zika virus disease in San Andres, Colombia.

zika_sanandres_2015

Format

A data frame with 101 rows and 2 columns

date

Date

cases

Daily incidence

The data on Zika virus disease incidence reported by Rojas et al. (2016) cover the period from September 2015 to January 2016, over which time a total of 928 cases were reported to health authorities of San Andres (population of 54,513). Suspected cases were confirmed by reverse transcription-polymerase chain reaction (RT-PCR) in the serum of acute cases within five days of symptom onset.

Source

Rojas et al. (2016)

References

Rojas, D. P., Dean, N. E., Yang, Y., Kenah, E., Quintero, J., Tomasi, S., ... Eyrolle-Guignot, D. (2016). The epidemiology and transmissibility of Zika virus in Girardot and San Andres island, Colombia, September 2015 to January 2016. Eurosurveillance, 21(28), 30283. http://doi.org/10.2807/1560-7917.ES.2016.21.28.30283

These data were provided under a Creative Commons Attribution (CC BY 4.0) Licence (http://creativecommons.org/licenses/by/4.0/).

Examples

## show first few days of Zika incidence
head(zika_sanandres_2015)

#>         date cases
#> 1 2015-09-06     1
#> 2 2015-09-07     1
#> 3 2015-09-08     1
#> 4 2015-09-13     1
#> 5 2015-09-15     4
#> 6 2015-09-19     1