Zika on the Yap Main Islands, Micronesia, 2007

These data describe weekly incidence of probable and confirmed cases of Zika virus on the Yap Main Islands, Micronesia.

zika_yap_2007

Format

A data frame with 29 rows and 3 columns

onset_date

Date

nr

Days after starting date

value

Number of cases per week

The data on weekly cases reported by Funk et al. (2016) cover the period from 2007-02-18 to 2007-09-02, over which time there were a total of 108 cases classified as probable (59) or confirmed (49) in a population of 7391. Cases were identified by a combination of prospective and retrospective surveillance at all health centres on Yap.

Source

Funk et al. (2016)

References

S. Funk, et al. 2016. Comparative Analysis of Dengue and Zika Outbreaks Reveals Differences by Setting and Virus. PLOS Neglected Tropical Diseases, 10(12), e0005173. http://doi.org/10.1371/journal.pntd.0005173

Examples

## show first few weeks of Zika incidence
head(zika_yap_2007)

#> # A tibble: 6 x 3
#>   onset_date    nr value
#>   <date>     <dbl> <dbl>
#> 1 2007-02-18     7     0
#> 2 2007-02-25    14     0
#> 3 2007-03-04    21     0
#> 4 2007-03-11    28     0
#> 5 2007-03-18    35     0
#> 6 2007-03-25    42     0