Dengue on the Yap Main Islands, Micronesia, 2011

These data describe incidence of clincal cases of Dengue on the Yap Main Islands, Micronesia.

dengue_yap_2011

Format

A data frame with 185 rows and 3 columns

onset_date

Date

nr

Days after starting date

value

Number of cases

The data on Dengue incidence reported by Funk et al. (2016) cover the period from 2011-07-07 to 2012-04-12, over which time a total of 978 cases were reported among 7391 residents. Suspected Dengue cases were identified by the Yap Department of Health, using the WHO (2009) case definition. A small proportion of cases (9 time series for the Yap Main Islands. Dengue virus serotype 2 was confirmed by reverse transcriptase polymerase chain reaction by the CDC Dengue Branch, Puerto Rico.

Source

Funk et al. (2016)

References

S. Funk, et al. 2016. Comparative Analysis of Dengue and Zika Outbreaks Reveals Differences by Setting and Virus. PLOS Neglected Tropical Diseases, 10(12), e0005173. http://doi.org/10.1371/journal.pntd.0005173

Examples

## show first few weeks of Dengue incidence
head(dengue_yap_2011)

#> # A tibble: 6 × 3
#>   onset_date    nr value
#>       <date> <dbl> <dbl>
#> 1 2011-07-07     7     0
#> 2 2011-07-14    14     0
#> 3 2011-07-21    21     0
#> 4 2011-07-28    28     0
#> 5 2011-08-04    35     0
#> 6 2011-08-11    42     0