Simulated Ebola outbreak

This simulated outbreak of Ebola Virus Disease matches some key properties of the West African Ebola outbreak of 2014-2015. Specifically, care was taken to use realistic delays (incubation period, serial interval, time to hospitalisation, etc.) and reproduction number (see references).

ebola_sim

ebola_sim_clean

Format

An object of class list of length 2.

Details

This dataset is used for teaching purposes during Imperial College's short course on infectious disease modelling. The exercise aims to simulate the response to an Ebola outbreak taking place in a single large city, and evaluate the impact of an intervention (increased bed capacity).

Note that to ensure realism, some errors have been introduced in this dataset. These can be identified as negative incubation periods (delay from infection to onset of symptoms). See example for a simple way to identify these cases. The dataset ebola_sim_clean is the same dataset, only dates of infection and onset have been set to 'NA'.

References

WHO Ebola Response Team. 2014. Ebola virus disease in West Africa--the first 9 months of the epidemic and forward projections. The New England journal of medicine 371:1481–1495.

WHO Ebola Response Team, J. Agua-Agum, A. Ariyarajah, B. Aylward, I. M. Blake, R. Brennan, A. Cori, C. A. Donnelly, I. Dorigatti, C. Dye, T. Eckmanns, N. M. Ferguson, P. Formenty, C. Fraser, E. Garcia, T. Garske, W. Hinsley, D. Holmes, S. Hugonnet, S. Iyengar, T. Jombart, R. Krishnan, S. Meijers, H. L. Mills, Y. Mohamed, G. Nedjati-Gilani, E. Newton, P. Nouvellet, L. Pelletier, D. Perkins, S. Riley, M. Sagrado, J. Schnitzler, D. Schumacher, A. Shah, M. D. Van Kerkhove, O. Varsaneux, and N. Wijekoon Kannangarage. 2015. West African Ebola epidemic after one year--slowing but not yet under control. The New England journal of medicine 372:584–587.

Examples

## identify mistakes in data entry (negative incubation period)
mistakes <- which(ebola_sim$linelist$date_of_onset <= ebola_sim$linelist$date_of_infection)
mistakes

#>  [1]   46   63  110  219  326  364  385  469  694  708  721  728  748  769  919
#> [16]  998 1097 1119 1216 1267 1305 1395 1469 1514 1597 1663 1692 1744 1766 1808
#> [31] 1959 2133 2241 2315 2855 2859 2941 3306 3642 3659 3699 3704 3730 3754 3787
#> [46] 3946 3994 4203 4262 4273 4686 4772 4881 4916 4993 5031 5097 5311 5815

ebola_sim$linelist[mistakes, ]

#>      case_id generation date_of_infection date_of_onset date_of_hospitalisation
#> 46    3f1aaf          4        2014-05-18    2014-05-18              2014-05-25
#> 63    ce9c02          5        2014-05-27    2014-05-27              2014-05-29
#> 110   700e84          6        2014-06-10    2014-06-10              2014-06-16
#> 219   8bd1e8          8        2014-07-10    2014-07-10              2014-07-11
#> 326   085bd1          8        2014-07-14    2014-07-14              2014-07-27
#> 364   5dc1e2          6        2014-07-29    2014-07-29              2014-07-31
#> 385   9c1ae9          9        2014-08-01    2014-08-01              2014-08-02
#> 469   6be0c7         11        2014-08-05    2014-08-05              2014-08-08
#> 694   78e5b6         10        2014-08-14    2014-08-14              2014-08-24
#> 708   228e8b         11        2014-08-21    2014-08-21              2014-08-24
#> 721   899d6c         10        2014-08-20    2014-08-20              2014-08-25
#> 728   2be2fa         13        2014-08-22    2014-08-22              2014-08-25
#> 748   137b57         10        2014-08-23    2014-08-23              2014-08-26
#> 769   9f2e47         10        2014-08-23    2014-08-23              2014-08-27
#> 919   9bdd46         10        2014-08-30    2014-08-30              2014-09-03
#> 998   2816f0         13        2014-09-03    2014-09-03              2014-09-05
#> 1097  1c1e91         14        2014-09-05    2014-09-05              2014-09-09
#> 1119  f461ef         12        2014-08-31    2014-08-31              2014-09-10
#> 1216  1cc5e0         13        2014-09-11    2014-09-11              2014-09-12
#> 1267  dfe4f5         12        2014-09-09    2014-09-09              2014-09-14
#> 1305  aee2e3         13        2014-09-12    2014-09-12              2014-09-15
#> 1395  34f7ac         12        2014-09-15    2014-09-15              2014-09-17
#> 1469  f90b8b         12        2014-09-18    2014-09-18              2014-09-18
#> 1514  dc0798         16        2014-09-19    2014-09-19              2014-09-19
#> 1597  898509         14        2014-09-20    2014-09-20              2014-09-21
#> 1663  1306c1         17        2014-09-22    2014-09-22              2014-09-22
#> 1692  7e1951         14        2014-09-21    2014-09-21              2014-09-23
#> 1744  893b3a         20        2014-09-22    2014-09-22              2014-09-24
#> 1766  1d1c7d         15        2014-09-21    2014-09-21              2014-09-25
#> 1808  922899         11        2014-09-20    2014-09-20              2014-09-26
#> 1959  737799         17        2014-09-26    2014-09-26              2014-09-29
#> 2133  01e6cd         14        2014-10-02    2014-10-02              2014-10-02
#> 2241  e8a743         12        2014-10-04    2014-10-04              2014-10-05
#> 2315  eeecde         15        2014-10-06    2014-10-06              2014-10-06
#> 2855  4e8443         13        2014-10-19    2014-10-19              2014-10-20
#> 2859  5993d0         20        2014-10-19    2014-10-19              2014-10-20
#> 2941  657aa1         11        2014-10-20    2014-10-20              2014-10-23
#> 3306  b1b620         21        2014-11-01    2014-11-01              2014-11-01
#> 3642  baa1e5         19        2014-11-09    2014-11-09              2014-11-12
#> 3659  e41a89         17        2014-11-07    2014-11-07              2014-11-13
#> 3699  8d591b         16        2014-11-13    2014-11-13              2014-11-14
#> 3704  2cbe04         19        2014-11-13    2014-11-13              2014-11-14
#> 3730  f1551b         18        2014-11-14    2014-11-14              2014-11-15
#> 3754  8bd6eb         15        2014-11-15    2014-11-15              2014-11-16
#> 3787  00031d         18        2014-11-17    2014-11-17              2014-11-17
#> 3946  03d33a         20        2014-11-23    2014-11-23              2014-11-25
#> 3994  32af98         20        2014-11-25    2014-11-25              2014-11-26
#> 4204  9db989         21        2014-12-04    2014-12-04              2014-12-06
#> 4263  c1a29b         21        2014-12-06    2014-12-06              2014-12-09
#> 4274  ab2d3f         22        2014-12-08    2014-12-08              2014-12-09
#> 4687  7375d2         23        2015-01-01    2015-01-01              2015-01-02
#> 4773  b69258         19        2015-01-07    2015-01-07              2015-01-08
#> 4882  25e201         22        2015-01-13    2015-01-13              2015-01-16
#> 4917  1873c3         22        2015-01-17    2015-01-17              2015-01-18
#> 4994  ef70cd         21        2015-01-22    2015-01-22              2015-01-24
#> 5032  ca9ec0         26        2015-01-25    2015-01-25              2015-01-26
#> 5098  0ee1ae         20        2015-01-29    2015-01-29              2015-02-01
#> 5312  b51296         23        2015-02-19    2015-02-19              2015-02-20
#> 5816  87be9a         30        2015-04-18    2015-04-18              2015-04-19
#>      date_of_outcome outcome gender
#> 46        2014-05-25    <NA>      f
#> 63        2014-06-17   Death      m
#> 110       2014-06-28    <NA>      m
#> 219       2014-07-16    <NA>      m
#> 326       2014-07-30    <NA>      m
#> 364       2014-08-02   Death      f
#> 385       2014-08-23    <NA>      f
#> 469             <NA> Recover      m
#> 694       2014-08-27    <NA>      f
#> 708       2014-08-30    <NA>      m
#> 721       2014-08-26    <NA>      f
#> 728             <NA>    <NA>      m
#> 748       2014-09-21 Recover      m
#> 769       2014-09-03 Recover      f
#> 919       2014-09-11 Recover      m
#> 998       2014-09-19    <NA>      m
#> 1097      2014-09-10    <NA>      m
#> 1119      2014-08-22 Recover      m
#> 1216      2014-09-22   Death      f
#> 1267      2014-09-26   Death      m
#> 1305      2014-09-27 Recover      f
#> 1395      2014-09-23 Recover      f
#> 1469      2014-09-13 Recover      m
#> 1514      2014-10-17    <NA>      f
#> 1597      2014-10-12 Recover      f
#> 1663      2014-09-29 Recover      f
#> 1692      2014-10-01 Recover      m
#> 1744      2014-10-04    <NA>      m
#> 1766      2014-10-02   Death      m
#> 1808      2014-10-01   Death      m
#> 1959      2014-10-11 Recover      m
#> 2133      2014-10-11    <NA>      f
#> 2241      2014-09-29   Death      f
#> 2315      2014-10-14   Death      f
#> 2855      2014-10-27 Recover      m
#> 2859      2014-11-01    <NA>      m
#> 2941      2014-11-06   Death      m
#> 3306      2014-12-08 Recover      m
#> 3642      2014-11-19    <NA>      f
#> 3659            <NA> Recover      f
#> 3699            <NA>    <NA>      f
#> 3704      2014-11-15   Death      f
#> 3730            <NA> Recover      f
#> 3754      2014-12-07   Death      f
#> 3787      2014-11-24   Death      f
#> 3946      2014-12-13 Recover      f
#> 3994      2014-12-08 Recover      m
#> 4204      2014-12-12   Death      m
#> 4263      2014-12-13 Recover      m
#> 4274      2014-12-16   Death      f
#> 4687      2015-01-03   Death      m
#> 4773            <NA>   Death      m
#> 4882            <NA>    <NA>      m
#> 4917      2015-01-29   Death      f
#> 4994      2015-01-25   Death      f
#> 5032      2015-02-05    <NA>      f
#> 5098      2015-02-08    <NA>      m
#> 5312      2015-02-25    <NA>      m
#> 5816      2015-05-11 Recover      f
#>                                          hospital       lon      lat
#> 46                             Connaught Hospital -13.24655 8.484859
#> 63                             Connaught Hospital -13.22437 8.470248
#> 110                            Connaught Hospital -13.26127 8.463422
#> 219                                          <NA> -13.24826 8.470268
#> 326                            Connaught Hospital -13.25968 8.464691
#> 364                             Military Hospital -13.24803 8.464861
#> 385                               Rokupa Hospital -13.26582 8.458813
#> 469                                         other -13.21285 8.470534
#> 694                            Connaught Hospital -13.23718 8.470879
#> 708                            Connaught Hospital -13.22360 8.470980
#> 721                                          <NA> -13.23917 8.475379
#> 728                            Connaught Hospital -13.26205 8.463343
#> 748                            Connaught Hospital -13.26271 8.481300
#> 769                            Connaught Hospital -13.21363 8.456702
#> 919                                          <NA> -13.22648 8.485700
#> 998                            Connaught Hospital -13.21981 8.451784
#> 1097                                        other -13.26476 8.460112
#> 1119                           Connaught Hospital -13.21816 8.479614
#> 1216                           Connaught Hospital -13.21869 8.484936
#> 1267                                        other -13.23263 8.467949
#> 1305                           Connaught Hospital -13.21060 8.479839
#> 1395                           Connaught Hospital -13.20947 8.450530
#> 1469                              Rokupa Hospital -13.26562 8.468240
#> 1514                           Connaught Hospital -13.24065 8.479631
#> 1597                                        other -13.24704 8.465761
#> 1663                           Connaught Hospital -13.22336 8.483234
#> 1692                                        other -13.21925 8.483758
#> 1744                           Connaught Hospital -13.23310 8.466730
#> 1766                                         <NA> -13.21249 8.478603
#> 1808                                         <NA> -13.22741 8.473975
#> 1959                            Mitylira Hospital -13.25798 8.483630
#> 2133                            Military Hospital -13.21157 8.467951
#> 2241                            Military Hospital -13.26574 8.449613
#> 2315                           Connaught Hospital -13.26144 8.465503
#> 2855                                        other -13.26615 8.475470
#> 2859                              Rokupa Hospital -13.25292 8.456837
#> 2941                            Mitylira Hospital -13.26967 8.462541
#> 3306                           Connaught Hospital -13.21515 8.487530
#> 3642                                         <NA> -13.22693 8.484813
#> 3659                           Connaught Hospital -13.25436 8.456684
#> 3699                            Military Hospital -13.25993 8.454327
#> 3704 Princess Christian Maternity Hospital (PCMH) -13.26583 8.462764
#> 3730                                        other -13.26714 8.476049
#> 3754                            Military Hospital -13.22479 8.471030
#> 3787                                         <NA> -13.24831 8.486500
#> 3946                           Connaught Hospital -13.23711 8.476159
#> 3994                           Connaught Hospital -13.26369 8.479313
#> 4204                                         <NA> -13.21466 8.488438
#> 4263                                        other -13.25438 8.459362
#> 4274                           Connaught Hospital -13.21589 8.462556
#> 4687                                         <NA> -13.25850 8.455848
#> 4773 Princess Christian Maternity Hospital (PCMH) -13.26536 8.472692
#> 4882                                         <NA> -13.25422 8.474586
#> 4917                                         <NA> -13.25600 8.481602
#> 4994                           Connaught Hospital -13.22143 8.462763
#> 5032                                         <NA> -13.21703 8.461714
#> 5098                                         <NA> -13.26896 8.462674
#> 5312                           Connaught Hospital -13.26148 8.465018
#> 5816                                        other -13.21062 8.451409

## check that ebola_sim_clean is identical after removing mistakes
identical(ebola_sim_clean$linelist, ebola_sim$linelist[-mistakes, ])

#> [1] FALSE