measles_hagelloch_1861 {outbreaks}	R Documentation

Measles in Hagelloch, Germany, 1861

Description

These data comprise of 188 cases of measles among children in the German city of Hagelloch, 1861. The data were originally collected by Dr. Albert Pfeilsticker (1863) and augmented and re-analysed by Dr. Heike Oesterle (1992).

Usage

measles_hagelloch_1861

Format

A data frame with 188 rows and 12 columns

case_ID

Case ID number

infector

Number of patient who is the putative source of infection

date_of_prodrome

Date of onset of prodromal symptoms

date_of_rash

Date of onset of rash

date_of_death

Date of death (NA implies recovered)

age

Age in years (fractions ignored)

gender

Gender of the individual (factor: f, m)

family_ID

Family ID number

class

School class (factor: 0, preschool; 1, 1st class; 2, 2nd class )

complications

Complications (factor: no, yes)

x_loc

x coordinate of house (in metres). Scaling in metres is obtained by multiplying the original coordinates by 2.5 (see details in Neal and Roberts (2004))

y_loc

y coordinate of house (in metres). See x_loc above.

Author(s)

This version of the data was formatted from hagelloch.df in the surveillance package, which in turn was provided by Niels Becker via Peter Neal. Formatting to fit in with the other datasets in the outbreaks package by Simon Frost (sdwfrost@gmail.com).

Source

Pfeilsticker (1863) and Oesterle (1992).

References

Pfeilsticker, A. 1863. Beiträge zur Pathologie der Masern mit besonderer Berücksichtigung der statistischen Verhältnisse, M.D. Thesis, Eberhard-Karls-Universität Tübingen. Available as http://www.archive.org/details/beitrgezurpatho00pfeigoog.

Oesterle, H. 1992. Statistische Reanalyse einer Masernepidemie 1861 in Hagelloch, M.D. Thesis, Eberhard-Karls-Universitäat Tübingen.

Neal, P. J. and Roberts, G. O. 2004. Statistical inference and model selection for the 1861 Hagelloch measles epidemic, Biostatistics 5(2):249-261.

Höhle M. 2007. surveillance: An R package for the monitoring of infectious diseases. Computational Statistics, 22:571-582.

Meyer, S., Held, L., & Höhle, M. 2017. Spatio-Temporal Analysis of Epidemic Phenomena Using the R Package surveillance. Journal of Statistical Software, 77(11), 1 - 55.

Examples

## show first few cases
head(measles_hagelloch_1861)

[Package outbreaks version 1.5.0 Index]