mers_korea_2015 {outbreaks}	R Documentation

Middle East respiratory syndrome in South Korea, 2015

Description

These datasets correspond to the initial information collected by the Epidemic Intelligence group at European Centre for Disease Prevention and Control (ECDC) during the first weeks of the outbreak of Middle East respiratory syndrome (MERS-CoV) outbreak (South Korea) in 2015. The data were used to follow the daily evolution of this outbreak using public information available.

Usage

mers_korea_2015

Format

A list of two dataframes:

$linelist: A dataframe of MERS-CoV cases and their attributes

id: Unique identifier

age: Age

age_class: Age using 10-year groups

sex: Sex

place_infect: Probable region of infection

reporting_ctry: Country reporting the case

loc_hosp: Local hospital name where the case was hospitalized

dt_onset: Date of onset of symptoms

dt_report: Date of reporting

week_report: Week number of date of reporting

dt_start_exp: Date of first probable exposure to another MERS Co-V case

dt_end_exp: Date of last probable exposure to another MERS Co-V case

dt_diag: Date of MERS Co-V diagnosis

outcome: Outcome (alive or dead)

dt_death: Date of death

$contacts: A dataframe describing the relationship between MERS Co-V cases

from: Unique identifier of the probably source patient

to: Unique identifier of the secondary case

exposure: Probable place of exposure

diff_dt_onset: Time in days between two successive cases

Details

This dataset is meant for teaching purposes; it represents neither the final outbreak investigation results nor a consolidated and complete description of the transmission chain.

Author(s)

Data collected by the European Centre for Disease Prevention and Control (Epidemic Intelligence and Response section, contact: Bertrand Sudre (bertrand.sudre@ecdc.europa.eu) and Kaja Kaasik Aaslav(Kaja.KaasikAaslav@ecdc.europa.eu). Transfer to R and documentation by Bertrand Sudre ( bertrand.sudre@ecdc.europa.eu).

References

More information on the intial stage of the outbreak in the following reference: Penttinen PM, Kaasik-Aaslav K, Friaux A, Donachie A, Sudre B, Amato-Gauci AJ, Memish ZA, Coulombier D. Taking stock of the first 133 MERS coronavirus cases globally–Is the epidemic changing? Euro Surveill. 2013 Sep 26;18(39). pii: 20596. PubMed PMID: 24094061.

Examples

## show the line list describing MERS Co-V cases and their attributes
head(mers_korea_2015$linelist)

## show the relationships between MERS Co-V cases
head(mers_korea_2015$contacts)

[Package outbreaks version 1.5.0 Index]