Simulated Varicella outbreak

Simulated Varicella outbreak

varicella_sim_berlin

Format

A data frame with 500 rows and 13 columns

sex

The gender of the simulated persons

ethnicity

Simulated ethnical origin

firstname

First names of the simulated persons

lastname

Last names of the simulated persons

age

Age of the simulated persons

center1

Name of the first center where the simulated persons stay

arrival1

Date of arrival at the first center

leave1

Date of departure at the first center

center2

Name of the second center where the simulated persons stay

arrival2

Date of arrival at the second center

leave2

Date of departure at the second center

onset

Date of onset of the disease

disease

Name of the disease

Background

This dataset is useful to compute incidence rates.

This dataset simulates an outbreak of varicella in german centers for foreigners. It is loosely based on the situation in 2015, when the numbers of foreigners seeking asylum exeded the available places in the center for foreigners. Varicella was the most frequent disease in these centers at that time. comparable with kindergartens and other shelters.

Description of infectious diseases in people seeking asylum in Germany in 2017 of Robert Koch-Institute, Berlin, Germany: https://www.rki.de/DE/Content/Gesundheitsmonitoring/Gesundheitsberichterstattung/GesundAZ/Content/A/Asylsuchende/Asylsuchende.html

The dataset was created by the package outbreakcreator https://github.com/jakobschumacher/outbreakcreator/.

Examples

head(varicella_sim_berlin)

#>      sex ethnicity     firstname  lastname age               center1   arrival1
#> 1 female    arabic       Saleema al-Ismael  16     Oranienburger Str 2015-05-06
#> 2   male  hispanic       Domonic Jaramillo  25 Platz der Luftbruecke 2015-01-19
#> 3 female    arabic       Haaniya  el-Mitri  34 Platz der Luftbruecke 2015-04-11
#> 4 female     asian         Erica   Tanghal  11 Platz der Luftbruecke 2015-01-20
#> 5   male     asian Fermin Joshua   Kikuchi  34 Platz der Luftbruecke 2015-06-15
#> 6   male    arabic        Muslim al-Moussa   5              Bizetstr 2014-11-18
#>       leave1               center2   arrival2     leave2      onset   disease
#> 1 2015-06-20              Bizetstr 2015-06-27 2015-08-19       <NA>      <NA>
#> 2 2015-01-21              Bizetstr 2015-01-30 2015-03-02       <NA>      <NA>
#> 3 2015-07-09 Platz der Luftbruecke 2015-07-18 2015-10-03 2015-07-06 varicella
#> 4 2015-03-30     Oranienburger Str 2015-04-08 2015-07-13 2015-08-20 varicella
#> 5 2015-08-17 Platz der Luftbruecke 2015-08-22 2015-11-01       <NA>      <NA>
#> 6 2015-02-03 Platz der Luftbruecke 2015-02-17 2015-04-27       <NA>      <NA>